# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plataforma de Cursos",
        "description": "API para gestionar cursos, instructores, lecciones, comentarios, calificaciones y favoritos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Lista cursos e instructores",
                "operationId": "listCourses",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "instructors_per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "instructors_cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de cursos e instructores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Crea un curso",
                "operationId": "storeCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "instructor_id",
                                    "title",
                                    "description",
                                    "price",
                                    "level"
                                ],
                                "properties": {
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel desde cero"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Curso completo para aprender Laravel 12 desde cero."
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "beginner"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Curso creado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                }
            }
        },
        "/api/courses/{course}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Obtiene un curso",
                "operationId": "showCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curso encontrado"
                    },
                    "404": {
                        "description": "Curso no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Actualiza un curso",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel avanzado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Contenido actualizado del curso."
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 59.99
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "advanced"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Curso actualizado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Elimina un curso",
                "operationId": "destroyCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Curso eliminado"
                    }
                }
            }
        },
        "/api/courses/{course}/average-rating": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Promedio de rating del curso",
                "operationId": "courseAverageRating",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promedio calculado"
                    }
                }
            }
        },
...
    },
...
}